refactor login file structure
