+++ frontend/src/app/app-routing.module.ts
@@ -1,9 +1,9 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from "@angular/router";
-import {LoginComponent} from "./login/login.component";
+import {LoginPageComponent} from "./login/login-page/login-page.component";
 
 const routes: Routes = [
-  {path: 'login', component: LoginComponent}
+  {path: 'login', component: LoginPageComponent}
 ];
 
 @NgModule({
