+++ frontend/src/app/app.module.ts
@@ -2,19 +2,19 @@ import {NgModule} from '@angular/core';
 import {BrowserModule} from '@angular/platform-browser';
 
 import {AppComponent} from './app.component';
-import {LoginFormComponent} from './login-form/login-form.component';
-import {RegisterFormComponent} from './register-form/register-form.component';
+import {LoginFormComponent} from './login/login-form/login-form.component';
+import {RegisterFormComponent} from './login/register-form/register-form.component';
 import {FormsModule} from "@angular/forms";
 import {HttpClientModule} from "@angular/common/http";
-import { LoginComponent } from './login/login.component';
-import { AppRoutingModule } from './app-routing.module';
+import {LoginPageComponent} from './login/login-page/login-page.component';
+import {AppRoutingModule} from './app-routing.module';
 
 @NgModule({
   declarations: [
     AppComponent,
     LoginFormComponent,
     RegisterFormComponent,
-    LoginComponent
+    LoginPageComponent
   ],
   imports: [
     BrowserModule,
