+++ frontend/src/app/login/login-page/login-page.component.html
@@ -1,2 +1,2 @@
-<app-login-form></app-login-form>
 <app-register-form></app-register-form>
+<app-login-form></app-login-form>
